# internal/infrastructure/http/handlers/api.go
// Package handlers provides HTTP handlers for the REST API
package handlers

import (
	"encoding/json"
	"net/http"
	"time"

	"github.com/alchemorsel/v3/internal/ports/inbound"
	"go.uber.org/zap"
)

// APIHandlers handles REST API requests
type APIHandlers struct {
	recipeService inbound.RecipeService
	logger        *zap.Logger
}

// NewAPIHandlers creates a new API handlers instance
func NewAPIHandlers(
	recipeService inbound.RecipeService,
	logger *zap.Logger,
) *APIHandlers {
	return &APIHandlers{
		recipeService: recipeService,
		logger:        logger,
	}
}

// APIResponse represents a standard API response
type APIResponse struct {
	Success bool        `json:"success"`
	Data    interface{} `json:"data,omitempty"`
	Error   string      `json:"error,omitempty"`
	Message string      `json:"message,omitempty"`
}

// ListRecipes handles GET /api/v3/recipes
func (h *APIHandlers) ListRecipes(w http.ResponseWriter, r *http.Request) {
	// TODO: Implement recipe listing
	response := APIResponse{
		Success: true,
		Data:    []interface{}{},
		Message: "Recipes retrieved successfully",
	}

	h.writeJSON(w, http.StatusOK, response)
}

// CreateRecipe handles POST /api/v3/recipes
func (h *APIHandlers) CreateRecipe(w http.ResponseWriter, r *http.Request) {
	// TODO: Implement recipe creation
	response := APIResponse{
		Success: true,
		Message: "Recipe created successfully",
	}

	h.writeJSON(w, http.StatusCreated, response)
}

// GetRecipe handles GET /api/v3/recipes/{id}
func (h *APIHandlers) GetRecipe(w http.ResponseWriter, r *http.Request) {
	// TODO: Implement recipe retrieval
	response := APIResponse{
		Success: true,
		Data:    map[string]interface{}{"id": "1", "title": "Sample Recipe"},
		Message: "Recipe retrieved successfully",
	}

	h.writeJSON(w, http.StatusOK, response)
}

// UpdateRecipe handles PUT /api/v3/recipes/{id}
func (h *APIHandlers) UpdateRecipe(w http.ResponseWriter, r *http.Request) {
	// TODO: Implement recipe update
	response := APIResponse{
		Success: true,
		Message: "Recipe updated successfully",
	}

	h.writeJSON(w, http.StatusOK, response)
}

// DeleteRecipe handles DELETE /api/v3/recipes/{id}
func (h *APIHandlers) DeleteRecipe(w http.ResponseWriter, r *http.Request) {
	// TODO: Implement recipe deletion
	response := APIResponse{
		Success: true,
		Message: "Recipe deleted successfully",
	}

	h.writeJSON(w, http.StatusOK, response)
}

// LikeRecipe handles POST /api/v3/recipes/{id}/like
func (h *APIHandlers) LikeRecipe(w http.ResponseWriter, r *http.Request) {
	// TODO: Implement recipe like
	response := APIResponse{
		Success: true,
		Message: "Recipe liked successfully",
	}

	h.writeJSON(w, http.StatusOK, response)
}

// HealthCheck handles GET /api/v3/health
func (h *APIHandlers) HealthCheck(w http.ResponseWriter, r *http.Request) {
	response := APIResponse{
		Success: true,
		Data: map[string]interface{}{
			"status":    "healthy",
			"timestamp": time.Now().Unix(),
			"version":   "v3.0.0",
		},
		Message: "Service is healthy",
	}

	h.writeJSON(w, http.StatusOK, response)
}

// writeJSON writes a JSON response
// RateRecipe handles POST /api/v1/recipes/{id}/rating
func (h *APIHandlers) RateRecipe(w http.ResponseWriter, r *http.Request) {
	// TODO: Implement recipe rating
	response := APIResponse{
		Success: true,
		Message: "Recipe rated successfully",
	}

	h.writeJSON(w, http.StatusOK, response)
}

// GetUserRecipes handles GET /api/v1/users/{id}/recipes
func (h *APIHandlers) GetUserRecipes(w http.ResponseWriter, r *http.Request) {
	// TODO: Implement user recipes listing
	response := APIResponse{
		Success: true,
		Data:    []interface{}{},
		Message: "User recipes retrieved successfully",
	}

	h.writeJSON(w, http.StatusOK, response)
}

// GetUserFavorites handles GET /api/v1/users/{id}/favorites
func (h *APIHandlers) GetUserFavorites(w http.ResponseWriter, r *http.Request) {
	// TODO: Implement user favorites listing
	response := APIResponse{
		Success: true,
		Data:    []interface{}{},
		Message: "User favorites retrieved successfully",
	}

	h.writeJSON(w, http.StatusOK, response)
}

func (h *APIHandlers) writeJSON(w http.ResponseWriter, status int, data interface{}) {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(status)

	if err := json.NewEncoder(w).Encode(data); err != nil {
		h.logger.Error("Failed to encode JSON response", zap.Error(err))
		http.Error(w, "Internal Server Error", http.StatusInternalServerError)
	}
}